+++ test/integration/check_log/run_program.py
@@ -9,7 +9,6 @@
         test/integration/check_log/run_program.py
 
     Arguments:
-        None
 
 """
 
@@ -34,7 +33,6 @@ import check_log
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -46,10 +44,12 @@ class UnitTest(unittest.TestCase):
 
     Super-Class:  unittest.TestCase
 
-    Sub-Classes:  None
+    Sub-Classes:
 
     Methods:
         setUp -> Integration testing initilization.
+        test_search_or -> Test with or search clause.
+        test_search_and -> Test with and search clause.
         test_file_suppress -> Test file logs with standard out suppression.
         test_stdin_marker_empty -> Test with standard in and empty marker file.
         test_filter_data -> Test filter data.
@@ -71,7 +71,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -127,6 +126,56 @@ class UnitTest(unittest.TestCase):
             self.skipTest("Pre-conditions not met.")
 
         self.args_array = {"-f": [self.log_file1, self.log_file2]}
+        self.args_array2 = {"-f": [self.log_file1, self.log_file2],
+                            "-S": ["third", "line"], "-k": "and",
+                            "-o": self.test_out}
+        self.args_array3 = {"-f": [self.log_file1, self.log_file2],
+                            "-S": ["sixth", "new"], "-k": "or",
+                            "-o": self.test_out}
+
+    def test_search_or(self):
+
+        """Function:  test_search_or
+
+        Description:  Test with or search clause.
+
+        Arguments:
+
+        """
+
+        with gen_libs.no_std_out():
+            check_log.run_program(self.args_array3)
+
+        if os.path.isfile(self.test_out):
+            with open(self.test_out) as f_hdlr:
+                out_str = f_hdlr.readline().rstrip()
+
+            self.assertEqual(out_str, "This is the sixth line")
+
+        else:
+            self.assertTrue(False)
+
+    def test_search_and(self):
+
+        """Function:  test_search_and
+
+        Description:  Test with and search clause.
+
+        Arguments:
+
+        """
+
+        with gen_libs.no_std_out():
+            check_log.run_program(self.args_array2)
+
+        if os.path.isfile(self.test_out):
+            with open(self.test_out) as f_hdlr:
+                out_str = f_hdlr.readline().rstrip()
+
+            self.assertEqual(out_str, "This is the third line")
+
+        else:
+            self.assertTrue(False)
 
     def test_file_suppress(self):
 
@@ -135,7 +184,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test file logs with standard out suppression.
 
         Arguments:
-            None
 
         """
 
@@ -151,7 +199,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with standard in with an empty marker file.
 
         Arguments:
-            mock_atty -> Mock Ref:  check_log.sys.stdin
 
         """
 
@@ -178,7 +225,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test filter data.
 
         Arguments:
-            None
 
         """
 
@@ -203,7 +249,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test ignore messages.
 
         Arguments:
-            None
 
         """
 
@@ -228,7 +273,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test file log marker.
 
         Arguments:
-            None
 
         """
 
@@ -248,7 +292,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test file logs.
 
         Arguments:
-            None
 
         """
 
@@ -276,7 +319,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with standard in no data.
 
         Arguments:
-            mock_atty -> Mock Ref:  check_log.sys.stdin
 
         """
 
@@ -293,7 +335,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with standard in with marker file.
 
         Arguments:
-            mock_atty -> Mock Ref:  check_log.sys.stdin
 
         """
 
@@ -322,7 +363,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with standard in.
 
         Arguments:
-            mock_atty -> Mock Ref:  check_log.sys.stdin
 
         """
 
@@ -348,7 +388,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test clearing the marker file.
 
         Arguments:
-            None
 
         """
 
@@ -371,7 +410,6 @@ class UnitTest(unittest.TestCase):
         Description:  Clean up of integration testing.
 
         Arguments:
-            None
 
         """